// packages/frontend/src/generated/types/BlendingComponentCompositeRequest.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export interface BlendingComponentCompositeRequest {
  /**
   * @type string
   */
  amountUsed: string
  /**
   * @type string, uuid
   */
  sourceProductId: string
  /**
   * @type string, uuid
   */
  storageId: string
}