// packages/frontend/src/generated/types/CreateBlendingCompositeRequest.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import type { BlendingComponentCompositeRequest } from './BlendingComponentCompositeRequest.ts'
import type { BlendingResultCompositeRequest } from './BlendingResultCompositeRequest.ts'

export interface CreateBlendingCompositeRequest {
  /**
   * @type array
   */
  components: BlendingComponentCompositeRequest[]
  /**
   * @type string, uuid
   */
  contractorId: string
  /**
   * @type string, date-time
   */
  date: string
  /**
   * @type string
   */
  documentNumber: string
  /**
   * @type array
   */
  results: BlendingResultCompositeRequest[]
  /**
   * @type string, uuid
   */
  targetProductId: string
}