// packages/frontend/src/generated/zod/blendingComponentCompositeRequestSchema.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import type { ToZod } from '../.kubb/ToZod.ts'
import type { BlendingComponentCompositeRequest } from '../types/BlendingComponentCompositeRequest.ts'
import * as z from 'zod'

export const blendingComponentCompositeRequestSchema = z.object({
  amountUsed: z.string(),
  sourceProductId: z.string().uuid(),
  storageId: z.string().uuid(),
}) as unknown as ToZod<BlendingComponentCompositeRequest>