// packages/frontend/src/generated/zod/blendingResultCompositeRequestSchema.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import type { ToZod } from '../.kubb/ToZod.ts'
import type { BlendingResultCompositeRequest } from '../types/BlendingResultCompositeRequest.ts'
import * as z from 'zod'

export const blendingResultCompositeRequestSchema = z.object({
  producedAmount: z.string(),
  storageId: z.string().uuid(),
}) as unknown as ToZod<BlendingResultCompositeRequest>