// packages/frontend/src/generated/zod/createBlendingCompositeRequestSchema.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import type { ToZod } from '../.kubb/ToZod.ts'
import type { CreateBlendingCompositeRequest } from '../types/CreateBlendingCompositeRequest.ts'
import * as z from 'zod'
import { blendingComponentCompositeRequestSchema } from './blendingComponentCompositeRequestSchema.ts'
import { blendingResultCompositeRequestSchema } from './blendingResultCompositeRequestSchema.ts'

export const createBlendingCompositeRequestSchema = z.object({
  components: z.array(z.lazy(() => blendingComponentCompositeRequestSchema)),
  contractorId: z.string().uuid(),
  date: z.string().datetime(),
  documentNumber: z.string(),
  results: z.array(z.lazy(() => blendingResultCompositeRequestSchema)),
  targetProductId: z.string().uuid(),
}) as unknown as ToZod<CreateBlendingCompositeRequest>